Fix error in README and expand slightly
The only thing that was actually wrong is that `kissh` takes a path to
the public key, not the private one. But I think it's also helpful to
make as much copy/pastable as we can.

The extra spaces around the URL are to work around what appears to be a
bug in my VIM markdown syntax highlighting.