Allow capitals in usernames, and fix GH CI
PR #24 added a Github username with captials in, which is not a valid
linux username by default.

This adds the --force-badname to work around that.

It also addresses the issue that allowed the above PR to pass tests - we
were erroneous running CI tests against main, not the PR branch.

We need the local git to have a checkout of main, in order to test the
update.sh script, but we were checking it out *after* the branch
checkout, so we were actually always testing against main. Tests worked
locally as the checkouts were only in CI.